chore: avoid portfolio and market page rerendering on sidebar view change
